+++ src/index.ts
@@ -3,8 +3,8 @@
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
 
-import { Link } from './link';
-
-import { LinkProps } from './link/Link';
+import { Link, LinkProps } from './link';
+import { ButtonLink, ButtonLinkProps } from './button';
 
 export { Link, LinkProps };
+export { ButtonLink, ButtonLinkProps };