+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
-## v0.1.0
+## Unreleased
+
+- Add tests & coverage
+- Add es6 distribution
+- Add ButtonLink component
+
+## v0.1.0 - 20170108
 
 - Add Link component